+++ webdev/lib/src/command/serve_command.dart
@@ -59,6 +59,13 @@ refresh: Performs a full page refresh.
         help: 'Automatically launches your application in Chrome with the '
             'debug port open. Use $chromeDebugPortFlag to specify a specific '
             'port to attach to an already running chrome instance instead.')
+    ..addOption(userDataDirFlag,
+        defaultsTo: null,
+        help: 'Use with $launchInChromeFlag to specify user data directory '
+            'to pass to chrome. Will start chrome window logged into default '
+            'profile with enabled extensions. Use `auto` as a value to infer '
+            'the default directory for the current OS. '
+            'Note: only supported for Mac OS X and linux platforms.')
     ..addFlag(liveReloadFlag,
         negatable: false,
         help: 'Automatically refreshes the page after each successful build.\n'