+++ webdev/lib/src/serve/dev_workflow.dart
@@ -72,7 +72,11 @@ Future<Chrome> _startChrome(
   ];
   try {
     if (configuration.launchInChrome) {
-      return await Chrome.start(uris, port: configuration.chromeDebugPort);
+      var userDataDir = configuration.userDataDir == autoOption
+          ? autoDetectChromeUserDataDirectory()
+          : configuration.userDataDir;
+      return await Chrome.start(uris,
+          port: configuration.chromeDebugPort, userDataDir: userDataDir);
     } else if (configuration.chromeDebugPort != 0) {
       return await Chrome.fromExisting(configuration.chromeDebugPort);
     }